VIB-843: Lin API Overhaul (#354)
* VIB-843: Overhaul Init, SendFrameHeader; Add UpdateTxBuffer and GetSlaveConfig

* VIB-843: Review suggestions; Correct C-API UpdateTxBuffer

* VIB-843: Use UpdateTxBuffer in Lin Demos

* VIB-843: Get rid of util/functional in Cpp Demo

* VIB-843: Fix GoToSleep behavior

+++ SilKit/source/services/pubsub/DataMessageDatatypeUtils.hpp
@@ -45,7 +45,7 @@ struct SourceInfo
     }
     struct HashFunction
     {
-        size_t operator()(const SourceInfo& s) const { 
+        uint64_t operator()(const SourceInfo& s) const { 
             auto hMediaType = SilKit::Util::Hash::Hash(s.mediaType);
             auto hLabels = SilKit::Util::Hash::Hash(s.labels);
             return SilKit::Util::Hash::HashCombine(hMediaType, hLabels);
